@@ -12,18 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Threading.Tasks;
-using Energinet.DataHub.MessageArchive.EntryPoint.BlobServices;
-using Energinet.DataHub.MessageArchive.EntryPoint.Factories;
+using Energinet.DataHub.Core.App.Common.Diagnostics.HealthChecks;
+using Energinet.DataHub.Core.App.FunctionApp.Diagnostics.HealthChecks;
+using Energinet.DataHub.MessageArchive.Common;
+using Energinet.DataHub.MessageArchive.Common.SimpleInjector;
 using Energinet.DataHub.MessageArchive.EntryPoint.Functions;
-using Energinet.DataHub.MessageArchive.EntryPoint.Handlers;
-using Energinet.DataHub.MessageArchive.EntryPoint.Models;
-using Energinet.DataHub.MessageArchive.EntryPoint.Repository;
-using Energinet.DataHub.MessageArchive.EntryPoint.Repository.Containers;
-using Energinet.DataHub.MessageArchive.EntryPoint.SimpleInjector;
-using Microsoft.Azure.Cosmos;
-using Microsoft.Azure.Cosmos.Fluent;
+using Energinet.DataHub.MessageArchive.EntryPoint.Monitor;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -33,124 +27,29 @@ using Container = SimpleInjector.Container;
 
 namespace Energinet.DataHub.MessageArchive.EntryPoint
 {
-    internal sealed class Startup : IAsyncDisposable
+    internal sealed class Startup : StartupBase
     {
-        public Startup()
+        protected override void Configure(IServiceCollection services)
         {
-            Container = new Container();
-        }
-
-        public Container Container { get; }
-
-        public async ValueTask DisposeAsync()
-        {
-            await Container.DisposeAsync().ConfigureAwait(false);
-            GC.SuppressFinalize(this);
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            SwitchToSimpleInjector(services);
-
-            services.AddLogging();
-            services.AddSimpleInjector(Container, x =>
-            {
-                x.DisposeContainerWithServiceProvider = !true;
-                x.AddLogging();
-            });
-
-            // config
             var config = services.BuildServiceProvider().GetService<IConfiguration>();
-            Container.RegisterSingleton(() => config!);
 
-            // Add Application insights telemetry
-            services.SetupApplicationInsightTelemetry(config ?? throw new InvalidOperationException());
-
-            // TODO Can these two be combined ?
-            RegisterArchiveStorageReader(Container);
-            RegisterBlobReader(Container);
-
-            RegisterBlobArchive(Container);
-            RegisterCosmosStorageWriter(Container);
-
-            Container.Register<IBlobProcessingHandler, BlobProcessingHandler>(Lifestyle.Transient);
-            Container.Register<ArchiveSearchRequestListener>(Lifestyle.Scoped);
-            Container.Register<RequestResponseLogTriggerFunction>(Lifestyle.Scoped);
-            Container.Register<ArchiveDownloadRequestResponseLog>(Lifestyle.Scoped);
+            // Health check
+            services
+                .AddHealthChecks()
+                .AddLiveCheck()
+                .AddAzureBlobStorage(
+                    config.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONNECTION_STRING"),
+                    config.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONTAINER_NAME"))
+                .AddAzureBlobStorage(
+                    config.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONNECTION_STRING"),
+                    config.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_PROCESSED_CONTAINER_NAME"),
+                    name: "azureblob-container-processed")
+                .AddCosmosDb(
+                    config.GetValue<string>("COSMOS_MESSAGE_ARCHIVE_CONNECTION_STRING"),
+                    "message-archive");
         }
 
-        private static void RegisterArchiveStorageReader(Container container)
-        {
-            container.Register<IStorageStreamReader>(() =>
-            {
-                var configuration = container.GetService<IConfiguration>();
-
-                var connectionString = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONNECTION_STRING");
-                var containerName = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_PROCESSED_CONTAINER_NAME");
-
-                var factory = new StorageServiceClientFactory(connectionString);
-                var storageConfig = new StorageConfig(containerName);
-
-                return new BlobStorageStreamReader(factory, storageConfig);
-            });
-        }
-
-        private static void RegisterBlobReader(Container container)
-        {
-            container.Register<IBlobReader>(() =>
-            {
-                var configuration = container.GetService<IConfiguration>();
-
-                var connectionString = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONNECTION_STRING");
-                var containerName = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONTAINER_NAME");
-                return new BlobReader(connectionString, containerName);
-            });
-        }
-
-        private static void RegisterBlobArchive(Container container)
-        {
-            container.Register<IBlobArchive>(() =>
-            {
-                var configuration = container.GetService<IConfiguration>();
-
-                var connectionString = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONNECTION_STRING");
-                var fromContainerName = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_CONTAINER_NAME");
-                var toContainerName = configuration.GetValue<string>("STORAGE_MESSAGE_ARCHIVE_PROCESSED_CONTAINER_NAME");
-
-                return new BlobArchive(connectionString, fromContainerName, toContainerName);
-            });
-        }
-
-        private static void RegisterCosmosStorageWriter(Container container)
-        {
-            container.Register<IStorageWriter<CosmosRequestResponseLog>, ArchiveWriterRepository>();
-            container.RegisterSingleton(() => GetCosmosClient(container));
-            container.Register<IArchiveContainer, ArchiveContainer>(Lifestyle.Scoped);
-            container.Register<IArchiveReaderRepository, ArchiveReaderRepository>(Lifestyle.Scoped);
-        }
-
-        private static IArchiveCosmosClient GetCosmosClient(Container container)
-        {
-            var configuration = container.GetService<IConfiguration>();
-            var connectionString = configuration.GetValue<string>("COSMOS_MESSAGE_ARCHIVE_CONNECTION_STRING");
-
-            if (string.IsNullOrEmpty(connectionString))
-            {
-                throw new InvalidOperationException(
-                    "Please specify a valid CosmosDBConnection in the appSettings.json file or your Azure Functions Settings.");
-            }
-
-            var cosmosClient = new CosmosClientBuilder(connectionString)
-                .WithSerializerOptions(new CosmosSerializationOptions
-                {
-                    PropertyNamingPolicy = CosmosPropertyNamingPolicy.CamelCase,
-                })
-                .Build();
-
-            return new ArchiveCosmosClient(cosmosClient);
-        }
-
-        private static void SwitchToSimpleInjector(IServiceCollection services)
+        protected override void ConfigureSimpleInjector(IServiceCollection services)
         {
             var descriptor = new ServiceDescriptor(
                 typeof(IFunctionActivator),
@@ -158,6 +57,19 @@ namespace Energinet.DataHub.MessageArchive.EntryPoint
                 ServiceLifetime.Singleton);
 
             services.Replace(descriptor);
+
+            services.AddSimpleInjector(Container, x =>
+            {
+                x.DisposeContainerWithServiceProvider = false;
+                x.AddLogging();
+            });
+        }
+
+        protected override void Configure(Container container)
+        {
+            Container.Register<RequestResponseLogTriggerFunction>();
+            Container.Register<IHealthCheckEndpointHandler, HealthCheckEndpointHandler>(Lifestyle.Scoped);
+            Container.Register<HealthCheckEndpoint>(Lifestyle.Scoped);
         }
     }
 }
