@@ -29,7 +29,7 @@ namespace Energinet.DataHub.MessageArchive.IntegrationTests
 
                 if (File.Exists(localSettingsPath))
                 {
-                    var json = JsonSerializer.Deserialize<Dictionary<string, string>>(File.ReadAllBytes(localSettingsPath)) !;
+                    var json = JsonSerializer.Deserialize<Dictionary<string, string>>(File.ReadAllBytes(localSettingsPath))!;
                     if (json.TryGetValue("connectionString", out var cosmosConnectionString))
                     {
                         ConnectionString = cosmosConnectionString;
@@ -38,6 +38,12 @@ namespace Energinet.DataHub.MessageArchive.IntegrationTests
                     if (json.TryGetValue("databaseName", out var databaseName))
                     {
                         DatabaseName = databaseName;
+                    }
+
+                    if (json.TryGetValue("disableAzurite", out var disableAzuriteStr) &&
+                        bool.TryParse(disableAzuriteStr, out var disableAzurite))
+                    {
+                        DisableAzurite = disableAzurite;
                     }
                 }
             }
@@ -54,5 +60,7 @@ namespace Energinet.DataHub.MessageArchive.IntegrationTests
         internal static string ConnectionString { get; } = "AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
 
         internal static string DatabaseName { get; } = "message-archive";
+
+        internal static bool DisableAzurite { get; }
     }
 }
