@@ -13,8 +13,9 @@
 // limitations under the License.
 
 using System;
-using Energinet.DataHub.MessageArchive.EntryPoint.Models;
-using Energinet.DataHub.MessageArchive.EntryPoint.Validation;
+using System.Collections.Generic;
+using Energinet.DataHub.MessageArchive.Reader.Models;
+using Energinet.DataHub.MessageArchive.Reader.Validation;
 using Xunit;
 using Xunit.Categories;
 
@@ -31,10 +32,10 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
             searchCriteria.DateTimeFrom = string.Empty;
 
             // Act
-            var (valid, errorMessage) = SearchCriteriaValidation.Validate(searchCriteria);
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
 
             // Assert
-            Assert.False(valid);
+            Assert.False(result.Valid);
         }
 
         [Fact]
@@ -48,10 +49,10 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
             DateTimeOffset logCreatedDate = new DateTime(637782039340000000, DateTimeKind.Utc); // 2022-01-19 15:45:34
 
             // Act
-            var (valid, errorMessage) = SearchCriteriaValidation.Validate(searchCriteria);
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
 
             // Assert
-            Assert.True(valid);
+            Assert.True(result.Valid);
             Assert.True(searchCriteria.DateTimeFromParsed <= logCreatedDate);
             Assert.True(searchCriteria.DateTimeToParsed!.Value.Ticks >= logCreatedDate.Ticks);
         }
@@ -64,13 +65,11 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
             searchCriteria.DateTimeFrom = "2022-01-01";
             searchCriteria.DateTimeTo = "2022-01-19 11?00:01";
 
-            var logCreatedDate = new DateTime(637782039340000000, DateTimeKind.Utc); // 2022-01-19 15:45:34
-
             // Act
-            var (valid, errorMessage) = SearchCriteriaValidation.Validate(searchCriteria);
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
 
             // Assert
-            Assert.False(valid);
+            Assert.False(result.Valid);
             Assert.Null(searchCriteria.DateTimeToParsed);
             Assert.Null(searchCriteria.DateTimeToParsed);
         }
@@ -84,15 +83,17 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
             searchCriteria.DateTimeTo = "2022-01-19T15:45:34Z";
 
             DateTimeOffset logCreatedDate = new DateTime(637782039340000000, DateTimeKind.Utc); // 2022-01-19 15:45:34
+#pragma warning disable CA1305
             var logCreatedDateString = logCreatedDate.ToString();
+#pragma warning restore CA1305
 
             // Act
-            var (valid, errorMessage) = SearchCriteriaValidation.Validate(searchCriteria);
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
 
             var dateTimeParsedString = searchCriteria.DateTimeToParsed.ToString();
 
             // Assert
-            Assert.True(valid);
+            Assert.True(result.Valid);
             Assert.True(searchCriteria.DateTimeFromParsed <= logCreatedDate);
             Assert.True(searchCriteria.DateTimeToParsed >= logCreatedDate);
             Assert.True(dateTimeParsedString == logCreatedDateString);
@@ -109,13 +110,104 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
             DateTimeOffset logCreatedDate = new DateTime(637782039340000000, DateTimeKind.Utc); // 2022-01-19 15:45:34
 
             // Act
-            var (valid, errorMessage) = SearchCriteriaValidation.Validate(searchCriteria);
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
 
             // Assert
-            Assert.True(valid);
+            Assert.True(result.Valid);
             Assert.True(searchCriteria.DateTimeFromParsed <= logCreatedDate);
             Assert.True(searchCriteria.DateTimeToParsed >= logCreatedDate);
             Assert.True(searchCriteria.DateTimeToParsed!.Value.TimeOfDay == new TimeSpan(22, 59, 59));
+        }
+
+        [Fact]
+        public void Test_SearchParams_NoDateTime()
+        {
+            // Arrange
+            var searchCriteria = Create_ValidSearchCriteria();
+            searchCriteria.DateTimeFrom = null;
+            searchCriteria.DateTimeTo = null;
+
+            // Act
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
+
+            // Assert
+            Assert.False(result.Valid);
+        }
+
+        [Fact]
+        public void Test_SearchParams_ProcessType()
+        {
+            // Arrange
+            var processTypes = "d12";
+            var searchCriteria = Create_ValidSearchCriteria();
+            searchCriteria.DateTimeFrom = "2022-01-01T00:00:00.000+01:00";
+            searchCriteria.DateTimeTo = "2022-01-19T23:59:59.000+01:00";
+            searchCriteria.ProcessTypes = new List<string>() { processTypes };
+
+            // Act
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
+
+            // Assert
+            Assert.True(result.Valid);
+#pragma warning disable CA1308
+            Assert.Contains(processTypes.ToUpperInvariant(), searchCriteria.ProcessTypes);
+#pragma warning restore CA1308
+        }
+
+        [Fact]
+        public void Test_SearchParams_RsmName()
+        {
+            // Arrange
+            var rsmInputName = "Notifybillingmasterdata";
+            var searchCriteria = Create_ValidSearchCriteria();
+            searchCriteria.DateTimeFrom = "2022-01-01T00:00:00.000+01:00";
+            searchCriteria.DateTimeTo = "2022-01-19T23:59:59.000+01:00";
+            searchCriteria.RsmNames = new List<string>() { rsmInputName };
+
+            // Act
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
+
+            // Assert
+            Assert.True(result.Valid);
+#pragma warning disable CA1308
+            Assert.Contains(rsmInputName.ToLowerInvariant(), searchCriteria.RsmNames);
+#pragma warning restore CA1308
+        }
+
+        [Fact]
+        public void Test_SearchParams_IncludeRelated_NoMessageId()
+        {
+            // Arrange
+            var searchCriteria = Create_ValidSearchCriteria();
+            searchCriteria.DateTimeFrom = "2022-01-01T00:00:00.000+01:00";
+            searchCriteria.DateTimeTo = "2022-01-19T23:59:59.000+01:00";
+            searchCriteria.IncludeRelated = true;
+            searchCriteria.MessageId = null;
+
+            // Act
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
+
+            // Assert
+            Assert.True(result.Valid);
+            Assert.False(searchCriteria.IncludeRelated);
+        }
+
+        [Fact]
+        public void Test_SearchParams_IncludeRelated_IsTrue()
+        {
+            // Arrange
+            var searchCriteria = Create_ValidSearchCriteria();
+            searchCriteria.DateTimeFrom = "2022-01-01T00:00:00.000+01:00";
+            searchCriteria.DateTimeTo = "2022-01-19T23:59:59.000+01:00";
+            searchCriteria.IncludeRelated = true;
+            searchCriteria.MessageId = "1234";
+
+            // Act
+            var result = SearchCriteriaValidation.Validate(searchCriteria);
+
+            // Assert
+            Assert.True(result.Valid);
+            Assert.True(searchCriteria.IncludeRelated);
         }
 
         private static SearchCriteria Create_ValidSearchCriteria()
@@ -129,7 +221,7 @@ namespace Energinet.DataHub.MessageArchive.Tests.Validation
                 DateTimeTo = "2022-01-31",
                 FunctionName = "functionName",
                 InvocationId = "invocationId",
-                ProcessType = "processType",
+                ProcessTypes = new List<string>() { "processType" },
                 ReasonCode = "reasonCode",
                 IncludeRelated = false,
                 SenderId = "senderId",
